Support Elasticsearch with X-Pack Security
This allows the poller to supply a username/password combination for
using basic authentication to an Elasticsearch node. For this it now
uses Treq instead of Twisted's built-in `getPage`. Additionally it (for
now) disables TLS certificate verification as the default certificate
tool doesn't provide proper hostnames in the certificate and the node is
usually addressed as `localhost`.